iprutils: Add support for type 3 QAC command
The header size for the data returned by the QAC command has changed.  This
patch changes the ipr_array_query_data structure and the
__ipr_query_array_config function such that all the query data will be held
in the buf array.  Then the data get parsed to initialize the other fields
in the structure which includes the new u8 *data field.  This gets set up
to point to the beginning of the records in buf.

Signed-off-by: Wayne Boyer <wayneb@linux.vnet.ibm.com>