iprutils: fix serial number comparison
The find_multipath_vset routine used the ARRAY_SIZE calculation for the length
of the serial number in the memcmp call.  This calculation gives a length of 4
when the actual length is 8.  This can lead to false positives when looking
for the corresponding vset.  This in turn can lead to failures when trying to
delete arrays.

The fix is to use the IPR_SERIAL_NUM_LEN definition in the memcmp call.

Also, add displaying the serial number for arrays in the detailed information.

Signed-off-by: Wayne Boyer <wayneb@linux.vnet.ibm.com>