iprutils: increase buffer size for new adapter config data
With the maximum number of devices attached to one of the new
SIS64 adapters, the configuration data can reach 182896 bytes.

The maximum size is caluculated by adding up the following:
1 header at 8 bytes
240 array records at 192 bytes
512 vset records at 192 bytes
240 device records at 160 bytes
1 capability header at 8 bytes
4 capability records at 24 bytes

8 + 240 * 192 + 512 * 192 + 240 * 160 + 8 + 4 * 24 = 182896

The fix is to bump up the size of IPR_QAC_BUFFER_SIZE to 200000,
which solves the problem and gives some room for any future growth.
IPR_SIS32_QAC_BUFFER_SIZE is defined with the old value of 16000 for
the legacy adapters.

Signed-off-by: Wayne Boyer <wayneb@linux.vnet.ibm.com>