iprutils: bump up number of command status records
In testing with the Cubic-R adapter and Homerun storage drawer, we observed a
case where the call to ipr_query_command_status() from the get_status() routine
is returning more than 100 records.

The fix is to bump up the number of ipr_cmd_status_records to 250 for now.
Also, the num_records field is really 16 bits.

Signed-off-by: Wayne Boyer <wayneb@linux.vnet.ibm.com>