iprutils: fix SES microcode image version on CLI
When updating the firmware of a SES device, the iprconfig CLI cannot get the
right version of the image file being used, causing the utility to show a wrong
version on the system log and report an error code even when the update actually
succeeded. The GUI update process works correctly.

This patch fixes this problem by calling the appropriate procedure when updating
SES devices to get the version from the image file.

Signed-off-by: Kleber Sacilotto de Souza <klebers@linux.vnet.ibm.com>