iprutils: add a flag to raid create command for known zeroed drives
When RAID arrays are created using the command line interface the array can
take a long time to be created since the state of the disks are not known.
This time can be shortened if the drives are formatted beforehand, however,
on the command line interface the information that the drives are known to be
zeroed can not be kept by the utility. This patch adds a "-z" flag to the
raid-create CLI command to inform the utility that the drives are known to be
zeroed.

Signed-off-by: Kleber Sacilotto de Souza <klebers@linux.vnet.ibm.com>