iprconfig: get location code when SES page 4 is unavailable
If SES page 4 is not supported on a system, the physical location
code of a drive can still be determined using information from page 7.

Page 7 specifies the location code from the planar down to the device
(i.e. "-Pn[-Cn]-Dn"). On PAPR platforms, the planar location label is
preceded only by the unit label (i.e. "Uxxxx.xxx.xxxxxxx-"). For
convenience, we determine the unit label by copying it from the
location label of the drive's I/O adapter. If the platform doesn't
support SES page 7 either, no location information is printed.

Note: the buffer length is 1 byte larger than it needs to be in
order to compensate for the fact that
ipr_strncpy_0(dest, src, length) actually writes length+1 bytes to
dest.

Signed-off-by: Jesse Larrew <jlarrew@linux.vnet.ibm.com>