iprutils: fix add_string_to_body() for non-interactive interface
If add_string_to_body() is called when a non-interactive command is being run
it will fail to get the number of columns from the ncurses API. This issue
is currently affecting the reclaim cache CLI commands.

This patch fix this issue by using the TIOCGWINSZ ioctl to get the terminal
window size, and if it fails (e.g. on some console environments) the column
size is set to the default value of 80.

Signed-off-by: Kleber Sacilotto de Souza <klebers@linux.vnet.ibm.com>