iprutils: fix wrong secondary status
In some scenarios the adapters can be wrongly identified as secondaries,
despite the fact that they are not connected to other adapters.

The function get_dual_ioa_state() relies on the information from the
cur_state field of the ipr_dual_ioa_entry struct to set the is_secondary
flag, even though the query multiple adapter status returns 0 as the
number of multiple adapter entries. In this case, if we use the
ioa_entry struct we can wrongly change the is_secondary flag.

This patch fix this issue by not changing the default multi adapter
configuration when the number of multiple adapter entries is 0.

Signed-off-by: Kleber Sacilotto de Souza <klebers@linux.vnet.ibm.com>