iprutils: fix column width
After commit f2f69447 (iprutils: suspend/resume utility for BlueHawk),
the iprutils output of some screens had the columns header "PCI/SCSI
Location" and "Platform Location" shortened by one character, but the
contents of the columns were not changed, so the contents of the
following columns are not aligned with the header.

This patch fixes the issue by changing the columns content width, so
columns are aligned with their headers.

Signed-off-by: Kleber Sacilotto de Souza <klebers@linux.vnet.ibm.com>