iprutils: Adds three new CLI APIs: query-array, query-device, and query-location.
query-array: Takes as an argument a platform location code. If the disk
at that location is in a RAID array, the device name of the RAID array is
returned.

query-device: Takes as an argument a platform location code. Prints
the device name of the device at that location.

query-location: Takes as an argument a device name. If the device specified is
a
disk, prints the physical location of the disk. If the device specified is a
single disk RAID 0 array, prints the physical location of the sole device in
the RAID 0 array.

Signed-off-by: Brian King <brking@linux.vnet.ibm.com>