iprutils: fixing daemons being stopped
This patch fixes an issue when the ipr daemons are stopped when
tool_init() function is called on iprlib.c. When failing to open
scsi_host class or read the scsi device model, the function called
exit_on_error(), which killed the daemons in the process. This patch
changes it to warning messages while performs 4 attempts to open
scsi_host class or read the scsi device model.

Signed-off-by: Daniel Kreling <kreling@linux.vnet.ibm.com>
Reviewed-by: Gabriel K. Bertazi <krisman@linux.vnet.ibm.com>
Reviewed-by: Brian King <brking@linux.vnet.ibm.com>
Signed-off-by: Brian King <Brian King brking@linux.vnet.ibm.com>