iprutils: Fix potential exploit when writing log file.
SUSE reported a potential exploitation that could lead iprconfig to
write to an evil symlink and damage protected files on the system.  This
patch fixes the issue by atomically creating and opening a temporary
file with an unpredictable name (mkstemp), so an attacker doesn't have
the opportunity to remove it and symlink the original log file to any
important file.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>