iprutils: simplifying and fixing specfile
Fix up the spec file to make the daemons to start at package installation and
update. Also, a redundant if hook was removed from %post and %preun blocks.
The if clauses were changed to verify the presence of systemctl, instead of a
macro defined path %{_unitdir}. Finally, all redirections to /dev/null were
removed in order to make installation more verbose.

Signed-off-by: Daniel Kreling <kreling@linux.vnet.ibm.com>
Reviewed-by : Gabriel K. Bertazi <krisman@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>