iprutils: Add warnings when creating arrays.
Raid creation may take some time when devices are not zeroed before
creation.  We make sure to handle the case when devices are not zeroed
more gratiously, allowing users to skip formatting, which might greatly
reduce the time required to complete array creation.

Also, add a warning if the user tries to create RAIDs 5 or 6 using HDDs
but they do not have an IOA write cache enabled.  This alerts the user
of performance degradation that might occur and asks for confirmation.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>