iprutils: Fail on libncurses if ncurses-devel is missing.
if package ncurses-devel is missing, iprutils build should fail.
Although, current configure script will search first for libform, a
library that is part of the ncurses package, and fail with error
"libform not found".  This patch changes the search order to look for
ncurses first, making this error display as "libncurses not found"
which is easier to understand and fix.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Reported-by: Breno Henrique leitao <leitao@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>