iprutils: Allow statically linking of iprconfig.
Statically linking is discouraged.  Although, there are situations in
which it is required.  This commit includes some required dependency
checks to enable building a statically linked version of iprconfig.  It
includes a new flag in the Autotools Build System,
'--enable-build-static' to support this option.  To keep static linked
binary to a minimum inside iprutils, only iprconfig has a static
version in this patch.

This also includes a new subpackage in the spec file to hold the static
binary.  To generate this subpackage, one must call rpmbuild passing
'--with static'.  The static version will be installed as
'iprconfig-static' and will not conflict with iprutils package.

Be aware that static building requires the static version of ncurses and
glibc packages.  Otherwise, iprconfig-static build will fail.

Since dependency order matters a lot when statically linking, this patch
depends heavily on my last fix, ("Fail on libncurses if ncurses-devel is
missing."), which is not upstream yet.  I made sure this one will not
even apply cleanly without it.  For testing purposes, make sure you have
that patch applied.

This was tested on F21, RHEL7x and SLES12.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>