iprutils: Improve error message regarding array rebuild rate.
Commit 15fdeaa65af2362d94e64ddd604b68f588495a05 ("Add
support to configurable rebuild rate in iprconfig.")  added
support to specify the array rebuild rate on command-line
interface of iprconfig. For a matter of usability, users can
set any value in the range 0-100, though internally the only
allowed values are in the range 0-15 - a conversion is made.
This patch only improves aesthetically the error message in
the case of the adapter does not allow the change of the
rebuild rate value - the return message used to show the
values in the range 0-15, but now it shows the values in
the same range of input, i.e, 0-100.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>