iprutils: Add support to set device write cache policy from ncurses.
Commit 0eb82ecf7470094aeeb215593ac8d21ff5444d80 ("Add command to set device
write cache policy.") added support to set the devices's write cache policy
on command-line interface of iprconfig. Here we add the same functionality
on the ncurses interface. As a side note, we added a call to the
check_current_config function after saving the configurations state on the
device - we believe that the call was missing, since another ncurses device
modification routines have this call after saving the device state
and we have found an erratic behaviour of our implementation without adding
the call.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>