iprutils: Add option to enable/disable initd scripts from configure script.
Allow users to enable initd scripts support in iprutils for distros that
support it.  One can run ../configure --with-initscripts to building the
init files.

This modifies the default behavior to not generate initd scripts by
default.  From now on, to generate them user must use --with-initscripts.

The default for --with-initscripts is to install scripts to
%{prefix}/etc/init.d.  That can be overriden using
--with-initscripts=<path>.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>