iprutils: Display adapter-wide information on vset Cache.
Add new field in show-details to report Sync Write Cache support for a
given IOA.  This will only be displayed when passing an IOA to command
'show-details'.

This field tells if the adapter has support to sync cache
enabled/implemented.  It does not mean that sync cache is enabled
for a specific device.  It has three states: Unsupported - in
case adapter doesn't support it; Enabled - if Sync Cache is
supported and is globally enabled for this adapter; Disabled - in case
it is supported but was disabled for the whole adapter.

The information is recovered from the device using the Query Cache
Parameters command.  It can return a new value now, indicating if Volume
Set Caching Mode Page is enabled.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>