iprutils: Configure cache during vset initialization.
Sync cache will be enabled at startup for every vset.  This add this
configuration to iprinit flow.

There is an issue with kernels that don't enable vset cache by default
for supported adapters, in which the adapter and the kernel disagree on
the current state of the WCE bit.  This happens since, in these kernels,
we enable vcache too late in the boot process, tricking scsi stack into
thinking WCE bit is disabled.  Since WCE bit is cached in the adapter,
to make things correct, we always reconfigure vcache in iprinit, even
when it is disabled by default.  This will force Linux to reload the WCE
information, fixing the synchronization issue.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Acked-by: Brian King <brking@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>