iprutils: Avoid conflict in the CLI parser.
In the case an iprconfig CLI command (-c <cmd>) accepts an argument
with the same name as one used by the iprconfig parser, the later will
interpret it as a global iprconfig argument, and not pass it to the
command specific parser.  For instance, a command line like this:

$ iprconfig -c raid-create -r 0 -c writeback ...

will fail, because the second '-c' is interpreted as an iprconfig
command, instead of an argument that must be passed to raid-create.

This patch raises the precedence of arguments in parsing, so anything
after the first '-c' will be interpreted as a command argument instead
of a global argument to iprconfig.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>