iprutils: Implement ipr dump formatting tool
This implements a new iprdumpfmt command that takes an iprdump.XX file as
input and strips out driver information. This includes driver version and type
and the driver trace in the correct order. It has a verbose option that prints
additional information for the driver trace, such as timestamps and SCSI
commands.

Signed-off-by: Heitor Ricardo Alves de Siqueira <halves@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>