ipr: Avoid inconsistent information about caching mode.
If we are on an old Bare Metal system (i.e. without the sync cache
kernel patches), but with a firmware recent enough to support page 0xC4
for Sync cache, iprconfig is tricked into thinking that a non-caching
adapter has some kind of cache available.  Therefore, it'll end up
printing misleading information, which confuses users and trigger lots
of bug reports.

This patch makes the bug reports stop arriving by dealing with this
case and fixing up the wrong information.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>