iprutils: Fix RAID Migration failure on Little Endian systems.
raid-migrate command fails on Little Endian machines if the stripe size
is not specified.  This happens because iprconfig will fallback to the
current stripe size of the RAID, loading it from the adapter without
doing proper endianness conversion.  This patch fixes the issue by
properly converting the data before using it.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>