iprutils: Remove system() calls from log functions
This commit rewrites some functions in iprconfig that are used to
display log information to the user. We now use standard C code instead
of system() calls and other OS utilities like grep. To view the
requested log info, the less pager is invoked with the environment
variable LESSSECURE set to 1 to prevent arbitrary command execution.

This commit also introduces a zlib depency to iprutils. This is needed
so that we can seamlessly work with compressed and uncompressed log files.

Changes since v1:
- moved invoke-pager() to iprconfig.c
- DEFAULT_EDITOR doesn't use an absolute path
- renamed some variables to make code more intelligible
- added a new entry to screen_status[] in iprconfig.h
- substituted return values with semantic literals from iprconfig.h
- added missing free() because of tmp_log
- fixed static build

Signed-off-by: Heitor Ricardo Alves de Siqueira <halves@linux.vnet.ibm.com>
Reviewed-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>