iprutis: Don't show RAID migration level for secondary devices.
When query-raid-levels-migrate is issued to a secondary (Non-optimized)
device on a dual adapter configuration, we send a Query Array
configuration with the Array Migrate Query option, which ends up
retrieving every RAID level supported by the adapter instead of only
returning the possible migration level candidates.

This patch blocks query-raid-levels-migrate for RAIDS that aren't ready
for migration, which avoids displaying impossible level candidates for
secondary devices and better complies with the behavior of the ncurses
interface.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>