configure.ac: use pow() instead of matherr() for libm check
In certain configurations of uClibc, the matherr() function may not be
provided by the C library, which makes the current configure.ac check
for libm fail. However, iprutils does not use matherr(), so using this
function for the test makes little sense.

This patch adjusts configure.ac to test for pow() instead, which is
actually used by iprutils, and more commonly available, including in
uClibc.

Signed-off-by: Bernd Kuhls <bernd.kuhls@t-online.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Reviewed-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>