iprutils: Improve documentation on set-array-rebuild-rate.
Explain in the man page that iprutils might not set the exact value when
configuring the rebuild rate.  Instead, it will convert the percentage
value provided by the user to a integer value in the range 0-15 before
writing to the adapter.  This means, that we only approximate to the
rate requested by the user and the return of 'query-array-rebuild-rate'
may not be the exact value previously passed to
'set-array-rebuild-rate'.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>