iprutils: Mark devices as 4K even if IOA doesn't support 4K AF DASD
If you put a JBOD 4K disk in an IOA that doesn't support 4K AF devices
it will work, but iprconfig will not recognize it as
IPR_BLK_DEV_CLASS_4K. In this case, when we try to format it for RAID,
iprconfig will try to set the block size as 512, which will fail the
format command silently.  This patch makes sure IPR_BLK_DEV_CLASS_4K is
set for every 4K device, despite the IOA supporting 4K AF devices or
not.  Then, the verification for it in format_for_raid():13852 will
actually work, and we will abort the format earlier with an error
message, instead of failing silently.

Setting IPR_BLK_DEV_CLASS_4K for every device should not be an issue,
since in every other place it is referenced, we also check for
dev->ioa->support_4k and ipr_is_af_dasd_device().

This also adds a '\n' at the end of the error message cited above, to
improve readability.

Signed-off-by: Gabriel Krisman Bertazi <krisman@linux.vnet.ibm.com>
Signed-off-by: Brian King <brking@linux.vnet.ibm.com>