iprutils: Fix find_multipath_jbod to never return itself
The find_multipath_jbod is intended to return the alternate path
for the same device in a multipath configuration. However, it was
simply returning the first device that matched the device_id which
may have been the same device as the function parameter. This fixes
some issues and errors that can be seen when formatting JBOD disks.

Signed-off-by: Brian King <brking@linux.vnet.ibm.com>