iprutils: Ensure device known zeroed state gets saved after format
This patch fixes a problem when formatting multiple disks, where either
another instance of iprconfig or iprinit or some similar tool ends up
calling evaluate_device before the issuing iprconfig does, resulting in
formatted devices that are not flagged as known zeroed. This adds some
locking around this function to ensure that does not happen.

Signed-off-by: Brian King <brking@linux.vnet.ibm.com>