iprutils: Fix driver unbind on format
When formatting a disk from JBOD to RAID, we unbind the disk
from the sd driver, including the remote path, in the case of
dual adapters, and flush unused multipaths. However, due to
a bug, this code wasn't actually working once the SCSI
host, bus, or target exceeded a single character.

Signed-off-by: Brian King <brking@linux.vnet.ibm.com>